Update home & detail UI
