+++ app/src/main/java/com/minotawr/gamers/ui/main/MainListDelegate.kt
@@ -1,0 +1,9 @@
+package com.minotawr.gamers.ui.main
+
+import com.minotawr.gamers.domain.model.Game
+
+interface MainListDelegate {
+    fun onItemClick(item: Game)
+
+    fun onFavorite(item: Game)
+}
