+++ app/src/main/res/layout/adapter_main.xml
@@ -18,106 +18,64 @@
             android:id="@+id/guideline"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.3" />
+            android:orientation="horizontal"
+            app:layout_constraintGuide_percent="0.7" />
 
         <ImageView
             android:id="@+id/img_game"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
+            android:layout_width="match_parent"
+            android:layout_height="200dp"
             android:scaleType="centerCrop"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintDimensionRatio="1:1"
-            app:layout_constraintEnd_toStartOf="@+id/guideline"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            tools:srcCompat="@android:color/black" />
+            tools:srcCompat="@drawable/limbo" />
+
+        <View
+            android:id="@+id/background"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:background="@color/colorBackgroundGray"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/guideline" />
 
         <TextView
             android:id="@+id/tv_title"
             style="@style/H2"
-            android:layout_width="0dp"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_margin="@dimen/dimen_12dp"
+            android:layout_marginStart="@dimen/dimen_12dp"
+            android:layout_marginTop="@dimen/dimen_12dp"
             android:fontFamily="@font/inter_medium"
-            android:textColor="#000"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="@+id/guideline"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="Title" />
-
-        <com.google.android.material.card.MaterialCardView
-            android:id="@+id/mcv_genre_1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            app:cardBackgroundColor="?attr/colorSecondary"
-            app:cardCornerRadius="6dp"
-            app:cardElevation="2dp"
-            app:contentPadding="@dimen/dimen_6dp"
+            android:textColor="@android:color/white"
             app:layout_constraintBottom_toTopOf="@+id/tv_rating"
-            app:layout_constraintStart_toStartOf="@id/tv_title"
-            app:layout_constraintTop_toBottomOf="@+id/tv_title"
-            app:layout_constraintVertical_bias="0"
-            app:layout_constraintVertical_chainStyle="spread_inside">
-
-            <TextView
-                android:id="@+id/tv_genre_1"
-                style="@style/B2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter_medium"
-                android:textColor="@android:color/black"
-                tools:text="Title" />
-
-        </com.google.android.material.card.MaterialCardView>
-
-        <com.google.android.material.card.MaterialCardView
-            android:id="@+id/mcv_genre_2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/dimen_8dp"
-            app:cardBackgroundColor="?attr/colorSecondary"
-            app:cardCornerRadius="6dp"
-            app:cardElevation="2dp"
-            app:contentPadding="@dimen/dimen_6dp"
-            app:layout_constraintBottom_toBottomOf="@+id/mcv_genre_1"
-            app:layout_constraintStart_toEndOf="@id/mcv_genre_1"
-            app:layout_constraintTop_toTopOf="@+id/mcv_genre_1">
-
-            <TextView
-                android:id="@+id/tv_genre_2"
-                style="@style/B2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter_medium"
-                android:textColor="@android:color/black"
-                tools:text="Title" />
-
-        </com.google.android.material.card.MaterialCardView>
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="@+id/background"
+            tools:text="Title" />
 
         <TextView
             android:id="@+id/tv_rating"
             style="@style/B3"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_margin="@dimen/dimen_12dp"
-            android:textColor="#AAA"
+            android:layout_marginBottom="@dimen/dimen_12dp"
+            android:textColor="@android:color/white"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toEndOf="@id/img_game"
-            app:layout_constraintTop_toBottomOf="@+id/mcv_genre_1"
-            app:layout_constraintVertical_chainStyle="spread"
+            app:layout_constraintStart_toStartOf="@id/tv_title"
+            app:layout_constraintTop_toBottomOf="@+id/tv_title"
             tools:text="Rated %d out of 5 stars" />
 
         <ImageView
             android:id="@+id/img_favorite"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_margin="@dimen/dimen_12dp"
             android:contentDescription="@string/content_description_floating_action_button_favorite"
             android:src="@drawable/ic_favorite_border"
-            app:layout_constraintBottom_toBottomOf="@+id/tv_rating"
-            app:layout_constraintEnd_toEndOf="@+id/tv_title"
-            app:layout_constraintTop_toTopOf="@+id/tv_rating" />
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:tint="@android:color/white" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
