+++ app/src/main/res/layout/skeleton_main.xml
@@ -14,98 +14,57 @@
         android:layout_height="wrap_content">
 
         <androidx.constraintlayout.widget.Guideline
-            android:id="@+id/guideline"
+            android:id="@+id/skeleton_guideline"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.3" />
+            android:orientation="horizontal"
+            app:layout_constraintGuide_percent="0.7" />
 
         <ImageView
-            android:id="@+id/img_game"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
+            android:id="@+id/skeleton_img_game"
+            android:layout_width="match_parent"
+            android:layout_height="200dp"
+            android:background="@android:color/darker_gray"
             android:scaleType="centerCrop"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintDimensionRatio="1:1"
-            app:layout_constraintEnd_toStartOf="@+id/guideline"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            android:background="@android:color/darker_gray"
             android:importantForAccessibility="no" />
 
+        <View
+            android:id="@+id/skeleton_background"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            android:background="@color/colorBackgroundShimmer"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/skeleton_guideline" />
+
         <TextView
-            android:id="@+id/tv_title"
+            android:id="@+id/skeleton_tv_title"
             style="@style/H2"
-            android:layout_width="0dp"
+            android:layout_width="@dimen/dimen_40dp"
             android:layout_height="wrap_content"
-            android:layout_margin="@dimen/dimen_12dp"
+            android:layout_marginStart="@dimen/dimen_12dp"
+            android:layout_marginTop="@dimen/dimen_12dp"
+            android:background="@android:color/darker_gray"
             android:fontFamily="@font/inter_medium"
-            android:textColor="#000"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="@+id/guideline"
-            app:layout_constraintTop_toTopOf="parent"
-            android:background="@android:color/darker_gray" />
-
-        <com.google.android.material.card.MaterialCardView
-            android:id="@+id/mcv_genre_1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="8dp"
-            app:cardBackgroundColor="@android:color/darker_gray"
-            app:cardCornerRadius="6dp"
-            app:cardElevation="2dp"
-            app:contentPadding="@dimen/dimen_6dp"
-            app:layout_constraintBottom_toTopOf="@+id/tv_rating"
-            app:layout_constraintStart_toStartOf="@id/tv_title"
-            app:layout_constraintTop_toBottomOf="@+id/tv_title"
-            app:layout_constraintVertical_bias="0"
-            app:layout_constraintVertical_chainStyle="spread_inside">
-
-            <TextView
-                android:id="@+id/tv_genre_1"
-                style="@style/B2"
-                android:layout_width="@dimen/dimen_32dp"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter_medium"
-                android:textColor="@android:color/black" />
-
-        </com.google.android.material.card.MaterialCardView>
-
-        <com.google.android.material.card.MaterialCardView
-            android:id="@+id/mcv_genre_2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/dimen_8dp"
-            app:cardBackgroundColor="@android:color/darker_gray"
-            app:cardCornerRadius="6dp"
-            app:cardElevation="2dp"
-            app:contentPadding="@dimen/dimen_6dp"
-            app:layout_constraintBottom_toBottomOf="@+id/mcv_genre_1"
-            app:layout_constraintStart_toEndOf="@id/mcv_genre_1"
-            app:layout_constraintTop_toTopOf="@+id/mcv_genre_1">
-
-            <TextView
-                android:id="@+id/tv_genre_2"
-                style="@style/B2"
-                android:layout_width="@dimen/dimen_32dp"
-                android:layout_height="wrap_content"
-                android:fontFamily="@font/inter_medium"
-                android:textColor="@android:color/black" />
-
-        </com.google.android.material.card.MaterialCardView>
+            android:textColor="@android:color/white"
+            app:layout_constraintBottom_toTopOf="@+id/skeleton_tv_rating"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="@+id/skeleton_background" />
 
         <TextView
-            android:id="@+id/tv_rating"
+            android:id="@+id/skeleton_tv_rating"
             style="@style/B3"
-            android:layout_width="@dimen/dimen_100dp"
+            android:layout_width="@dimen/dimen_120dp"
             android:layout_height="wrap_content"
             android:layout_marginBottom="@dimen/dimen_12dp"
             android:background="@android:color/darker_gray"
-            android:textColor="#AAA"
+            android:textColor="@android:color/white"
             app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="@+id/tv_title"
-            app:layout_constraintTop_toBottomOf="@+id/mcv_genre_1"
-            app:layout_constraintVertical_chainStyle="spread" />
+            app:layout_constraintStart_toStartOf="@id/skeleton_tv_title"
+            app:layout_constraintTop_toBottomOf="@+id/skeleton_tv_title" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
