+++ app/src/main/res/values/dimens.xml
@@ -8,4 +8,6 @@
     <dimen name="dimen_24dp">24dp</dimen>
     <dimen name="dimen_32dp">32dp</dimen>
     <dimen name="dimen_100dp">100dp</dimen>
+    <dimen name="dimen_40dp">40dp</dimen>
+    <dimen name="dimen_120dp">120dp</dimen>
 </resources>
